live updates

+++ src/modules/delete.ts
@@ -1,5 +1,5 @@
 import { CommandModule, ModuleType } from "discord-dbm";
-import { Message, TextChannel, DMChannel, GroupDMChannel } from "discord.js";
+import { Message, TextChannel, DMChannel, NewsChannel } from "discord.js";
 
 class Delete implements CommandModule {
     configuration = {
@@ -10,23 +10,26 @@ class Delete implements CommandModule {
     }
     
     async onCommand(command: string[], message: Message): Promise<void> {
-        this.searchForMessage(message.channel, message.client.user.id).then(id => {
+        const myId = message.guild?.me?.id;
+        if (myId === undefined) {
+            return;
+        }
+        this.searchForMessage(message.channel, myId).then(id => {
             if (id !== undefined) {
                 message.channel.bulkDelete([id]);
             }
         });
     }
 
-    async searchForMessage(channel: TextChannel | DMChannel | GroupDMChannel, clientUserId: string): Promise<string | undefined> {
-        const messages = await channel.fetchMessages({
+    async searchForMessage(channel: TextChannel | DMChannel | NewsChannel, clientUserId: string): Promise<string | undefined> {
+        const messages = await channel.messages.fetch({
             limit: 15
         });
         
         return messages
             .filter(value => value.author.id === clientUserId)
             .sort((v1, v2) => v1.createdTimestamp - v2.createdTimestamp)
-            .last()
-            .id;
+            .last()?.id;
     }
 }
